# apps/frontend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/itinerary-searches": {
      "post": {
        "tags": [
          "ItinerarySearch"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OptimizedTripRequest"
                  },
                  {
                    "$ref": "#/components/schemas/OrderedTripRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OptimizedTripRequest"
                  },
                  {
                    "$ref": "#/components/schemas/OrderedTripRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OptimizedTripRequest"
                  },
                  {
                    "$ref": "#/components/schemas/OrderedTripRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItinerarySearchSessionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItinerarySearchSessionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItinerarySearchSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/itinerary-searches/{searchId}": {
      "get": {
        "tags": [
          "ItinerarySearch"
        ],
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ranking",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Direct",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "OneStop",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "TwoPlusStops",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Airlines",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "BookingSources",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DepartureAirports",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ArrivalAirports",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MaxPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MaxDurationMinutes",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "DepartureTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ArrivalTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "BookingType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MaxBookingCount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "AllowAirportSwitches",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItinerarySearchSessionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItinerarySearchSessionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItinerarySearchSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ItinerarySearch"
        ],
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Search/{searchId}": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Direct",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "OneStop",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "TwoPlusStop",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Providers",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Airlines",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DepartureAirports",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ArrivalAirports",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MaxDuration",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "DepartureTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ArrivalTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ReturnDepartureTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ReturnArrivalTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OutboundLegId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ReturnLegId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "HasStopFilter",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IsPaginationRequested",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSessionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSessionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AirportGroupRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "airportCodes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BookingOption": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DestinationRequest": {
        "type": "object",
        "properties": {
          "group": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AirportGroupRequest"
              }
            ],
            "nullable": true
          },
          "stay": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StayRuleRequest"
              }
            ],
            "nullable": true
          },
          "airportContinuity": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ItineraryFilterMetadata": {
        "type": "object",
        "properties": {
          "airlines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryFilterOption"
            },
            "nullable": true
          },
          "bookingSources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryFilterOption"
            },
            "nullable": true
          },
          "departureAirports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryFilterOption"
            },
            "nullable": true
          },
          "arrivalAirports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryFilterOption"
            },
            "nullable": true
          },
          "minPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "maxPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "maxDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxBookingCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxAirportSwitches": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ItineraryFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ItineraryLeg": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "originAirport": {
            "type": "string",
            "nullable": true
          },
          "destinationAirport": {
            "type": "string",
            "nullable": true
          },
          "departureLocalTime": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalLocalTime": {
            "type": "string",
            "format": "date-time"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "stops": {
            "type": "integer",
            "format": "int32"
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItinerarySegment"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ItineraryPagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalResults": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ItineraryResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "bookingType": {
            "type": "string",
            "nullable": true
          },
          "destinationOrder": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryLeg"
            },
            "nullable": true
          },
          "stays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryStay"
            },
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "totalFlightDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "totalStops": {
            "type": "integer",
            "format": "int32"
          },
          "bookingCount": {
            "type": "integer",
            "format": "int32"
          },
          "airportSwitches": {
            "type": "integer",
            "format": "int32"
          },
          "bookingOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookingOption"
            },
            "nullable": true
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryWarning"
            },
            "nullable": true
          },
          "rankingBreakdown": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RankingBreakdown"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ItinerarySearchRequest": {
        "required": [
          "mode"
        ],
        "type": "object",
        "properties": {
          "mode": {
            "type": "string"
          },
          "adults": {
            "type": "integer",
            "format": "int32"
          },
          "cabinClass": {
            "type": "string",
            "nullable": true
          },
          "ranking": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "mode",
          "mapping": {
            "optimize": "#/components/schemas/OptimizedTripRequest",
            "ordered": "#/components/schemas/OrderedTripRequest"
          }
        }
      },
      "ItinerarySearchSessionResponse": {
        "type": "object",
        "properties": {
          "searchId": {
            "type": "string",
            "nullable": true
          },
          "mode": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "phase": {
            "type": "string",
            "nullable": true
          },
          "progress": {
            "type": "integer",
            "format": "int32"
          },
          "coverage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchCoverage"
              }
            ],
            "nullable": true
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryResult"
            },
            "nullable": true
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryWarning"
            },
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "pagination": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ItineraryPagination"
              }
            ],
            "nullable": true
          },
          "filters": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ItineraryFilterMetadata"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ItineraryStay": {
        "type": "object",
        "properties": {
          "destinationId": {
            "type": "string",
            "nullable": true
          },
          "arrivalDate": {
            "type": "string",
            "format": "date"
          },
          "departureDate": {
            "type": "string",
            "format": "date"
          },
          "nights": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ItineraryWarning": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OptimizedTripRequest": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ItinerarySearchRequest"
          }
        ],
        "properties": {
          "start": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AirportGroupRequest"
              }
            ],
            "nullable": true
          },
          "destinations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DestinationRequest"
            },
            "nullable": true
          },
          "endpointMode": {
            "type": "string",
            "nullable": true
          },
          "fixedEnd": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AirportGroupRequest"
              }
            ],
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "defaultAirportContinuity": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderedLegRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AirportGroupRequest"
              }
            ],
            "nullable": true
          },
          "to": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AirportGroupRequest"
              }
            ],
            "nullable": true
          },
          "departureDate": {
            "type": "string",
            "format": "date"
          },
          "airportContinuityWithPrevious": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderedTripRequest": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ItinerarySearchRequest"
          }
        ],
        "properties": {
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderedLegRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RankingBreakdown": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "format": "double"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "additionalFlightMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "totalStops": {
            "type": "integer",
            "format": "int32"
          },
          "additionalBookings": {
            "type": "integer",
            "format": "int32"
          },
          "airportSwitches": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SearchCoverage": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "nullable": true
          },
          "liveProviderCallsUsed": {
            "type": "integer",
            "format": "int32"
          },
          "providerCallLimit": {
            "type": "integer",
            "format": "int32"
          },
          "cacheHits": {
            "type": "integer",
            "format": "int32"
          },
          "candidateStatesEvaluated": {
            "type": "integer",
            "format": "int32"
          },
          "candidateStatesPruned": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SearchResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "nullable": true
          },
          "metadata": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchMetadata"
              }
            ],
            "nullable": true
          },
          "filters": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchFiltersMetadata"
              }
            ],
            "nullable": true
          },
          "pagination": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchPagination"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "isRoundTrip": {
            "type": "boolean"
          },
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResultLeg"
            },
            "nullable": true
          },
          "totalDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "priceOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResultPriceOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SearchSessionResponse": {
        "type": "object",
        "properties": {
          "searchId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "totalCombinations": {
            "type": "integer",
            "format": "int32"
          },
          "completedCombinations": {
            "type": "integer",
            "format": "int32"
          },
          "failedCombinations": {
            "type": "integer",
            "format": "int32"
          },
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchResponse"
              }
            ],
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "stagedResults": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SearchStagedResults"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SearchStagedResults": {
        "type": "object",
        "properties": {
          "outboundResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "nullable": true
          },
          "inboundResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "nullable": true
          },
          "roundTripResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StayRuleRequest": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "nullable": true
          },
          "nights": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  },
...
}